2009-03-12  Tao Wang  <dancefire@gmail.com>
	* cpp/include/openclas/serialization.hpp:
	* cpp/include/openclas/utf8_codecvt_facet.hpp:
	* cpp/src/unit_test/unit_test_serialization.hpp:
		Commented unused arguments.
		Add casting from enum to size_t.

+++ trunk/cpp/src/unit_test/unit_test_serialization.hpp
@@ -141,7 +141,7 @@ BOOST_AUTO_TEST_CASE( test_Serialization_load_tags_from_ctx )
 	const char* tag_name = "data/lexical.ctx";
 	Dictionary dict_core;
 	load_tags_from_ctx(dict_core, tag_name);
-	BOOST_CHECK_EQUAL( dict_core.tags().size(), WORD_TAG_SIZE );
+	BOOST_CHECK_EQUAL( dict_core.tags().size(), static_cast<size_t>(WORD_TAG_SIZE) );
 
 	//const char* nr_name = "data/nr.ctx";
 	//Dictionary dict_nr;
